+++ types/testutil/fixture/module_key.go
@@ -1,0 +1,8 @@
+package fixture
+
+type InvokerFactory func(callInfo CallInfo) (Invoker, error)
+
+type CallInfo struct {
+	Method string
+	Caller ModuleID
+}